Move raster wrapper classes into a new module
